requirements.txt
google-adk==0.5.0
requests
python-dotenv
